rf getVersionOnFooter
